ESQL: Move `randomLiteral`
This moves `randomLiteral` from the function testing infrastructure into
`EsqlTestUtils` so it can be used a little more widely. That'll make
other refactoring a little easier later.
