+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/session/EsqlConfigurationSerializationTests.java
@@ -25,7 +25,6 @@ import org.elasticsearch.test.AbstractWireSerializingTestCase;
 import org.elasticsearch.xpack.esql.Column;
 import org.elasticsearch.xpack.esql.action.ParseTables;
 import org.elasticsearch.xpack.esql.core.type.DataType;
-import org.elasticsearch.xpack.esql.expression.function.AbstractFunctionTestCase;
 import org.elasticsearch.xpack.esql.planner.PlannerUtils;
 import org.elasticsearch.xpack.esql.plugin.QueryPragmas;
 
@@ -34,6 +33,7 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
+import static org.elasticsearch.xpack.esql.EsqlTestUtils.randomLiteral;
 import static org.elasticsearch.xpack.esql.session.EsqlConfiguration.QUERY_COMPRESS_THRESHOLD_CHARS;
 
 public class EsqlConfigurationSerializationTests extends AbstractWireSerializingTestCase<EsqlConfiguration> {
@@ -117,7 +117,7 @@ public class EsqlConfigurationSerializationTests extends AbstractWireSerializing
                     )
                 ) {
                     for (int p = 0; p < positions; p++) {
-                        BlockUtils.appendValue(builder, AbstractFunctionTestCase.randomLiteral(dataType).value(), type);
+                        BlockUtils.appendValue(builder, randomLiteral(dataType).value(), type);
                     }
                     columns.put(name, new Column(dataType, builder.build()));
                 }
